+++ grafana-plugin/src/containers/AlertRules/AlertRules.tsx
@@ -804,8 +804,8 @@ class AlertRules extends React.Component<AlertRulesProps, AlertRulesState> {
         alertReceiveChannelStore.updateCounters();
         openNotification(
           <div>
-            Demo alert was generated. Find it in the
-            <PluginLink query={{ page: 'incidents' }}> "Incidents" </PluginLink>
+            Demo alert was generated. Find it on the
+            <PluginLink query={{ page: 'incidents' }}> "Alert Groups" </PluginLink>
             page and make sure it didn't freak out your colleagues 😉
           </div>
         );
@@ -821,8 +821,8 @@ class AlertRules extends React.Component<AlertRulesProps, AlertRulesState> {
       alertReceiveChannelStore.sendDemoAlertToParticularRoute(id).then(() => {
         openNotification(
           <div>
-            Demo alert was generated. Find it in the
-            <PluginLink query={{ page: 'incidents' }}> "Incidents" </PluginLink>
+            Demo alert was generated. Find it on the
+            <PluginLink query={{ page: 'incidents' }}> "Alert Groups" </PluginLink>
             page and make sure it didn't freak out your colleagues 😉
           </div>
         );